backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
crewai==0.11.2
langchain==0.0.344
requests==2.31.0
python-dotenv==1.0.0
websockets==11.0.3
aiofiles==23.2.1
numpy==1.26.1
Jinja2==3.1.2
ollama==0.1.5
huggingface-hub==0.19.4
